refactor: rename MeetingCard component and update imports for consistency
